Add discord-irc docker-compose configuration
This service connects the SMR discord server and IRC channel so that
messages posted to one are re-posted to the other by a bot.

No one really uses the IRC channel anymore, but I think it's good to
keep around in case something catastrophic happens to Discord.

See https://github.com/reactiflux/discord-irc.